rename install-copy-sqlite-esbuild to download-copy-sqlite-esbuild
