+++ extensions/vscode/scripts/download-copy-sqlite-esbuild.js
